Feat: RequireAuth / Get auth state from server and navigate user #51
